# api/openapi.yaml
openapi: 3.0.3
info:
  title: Kayran Platform API
  description: API для платформы водных активностей, проката, спорта и инклюзивных программ
  version: 0.1.0
  contact:
    name: Kayran Team
servers:
  - url: https://api.kayran.ru/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /users/{id}:
    get:
      tags: [Users]
      summary: Получить пользователя
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Данные пользователя

  # Boats
...
  /routes:
    get:
      tags: [Routes]
      summary: Список маршрутов
      parameters:
        - in: query
          name: difficulty
          schema: { type: string }
      responses:
        '200':
          description: Список маршрутов

    post:
      tags: [Routes]
      summary: Создать маршрут
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Маршрут создан
...
  /teams:
    get:
      tags: [Teams]
      summary: Список команд
      responses:
        '200':
          description: Список команд

    post:
      tags: [Teams]
      summary: Создать команду
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Команда создана
...
  /matches:
    get:
      tags: [Matches]
      summary: Список матчей
      parameters:
        - in: query
          name: tournament_id
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: Список матчей
  /matches/{id}/score:
    put:
      tags: [Matches]
      summary: Обновить счёт матча
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [score_a, score_b]
              properties:
                score_a: { type: integer }
                score_b: { type: integer }
      responses:
        '200':
          description: Счёт обновлён

  # Payments
  /payments:
    post:
      tags: [Payments]
      summary: Создать платёж
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Платёж создан

  # Inclusive
...
  /franchises:
    get:
      tags: [Franchises]
      summary: Список франшиз
      responses:
        '200':
          description: Список франшиз

    post:
      tags: [Franchises]
      summary: Создать франшизу
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Франшиза создана